Function ej_buscaminas::buscaminas::armar_tablero

fn armar_tablero(
    mapa: &[Vec<u8>], 
    bombas: &[Vec<u8>], 
    cant_filas: usize, 
    cant_columnas: usize
) -> Vec<Vec<u8>>

Arma el tablero final, con la cantidad de bombas adyacentes a todas las cuadrículas del mapa y las bombas según corresponda. Donde una cuadrícula contenga el símbolo de bomba, significa que en el mapa había una bomba allí. Donde una cuadrícula contenga un número, representa la cantidad de bombas adyacentes a la misma, inclusive el 0 -es decir, que dicha cuadrícula no tiene bombas adyacentes-. Recibe el mapa original en formato de vector de vectores de u8, el mapa con la información de la cantidad de bombas adyacentes en formato de vector de vectores de u8, y la cantidad de filas y de columnas que contiene el mapa original -y por ende todos los demás tableros construidos-. Entrega un tablero en formato de vector de vectores de u8 con la cantidad de bombas adyacentes a todas las cuadrículas y las bombas del mapa.